requirements.txt
pydub
yt_dlp
openai
python-dotenv
click
validators
pytest
tiktoken
